+++ main/packages/reactivizer/es/stream-core.js
@@ -78,7 +78,7 @@ export class ControllerCore {
         const dispatch = (...params) => {
             const action = this.createAction(type, params);
             this.actionUpstream.next(action);
-            return action.i;
+            return action;
         };
         this.dispatcher[type] = dispatch;
         return dispatch;
@@ -91,7 +91,7 @@ export class ControllerCore {
             const action = this.createAction(type, params);
             action.r = Array.isArray(metas) ? metas.map(m => m.i) : metas.i;
             this.actionUpstream.next(action);
-            return action.i;
+            return action;
         };
         this.dispatcherFor[type] = dispatch;
         return dispatch;